# openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "MeisterProPR.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/clients/{clientId}/finding-dismissals": {
      "get": {
        "tags": [
          "FindingDismissals"
        ],
        "summary": "Lists all finding dismissals for the specified client.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of dismissals.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingDismissalDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingDismissalDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingDismissalDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "FindingDismissals"
        ],
        "summary": "Creates a new finding dismissal for the specified client.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Dismissal details.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFindingDismissalRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFindingDismissalRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFindingDismissalRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dismissal created.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDismissalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDismissalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDismissalDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/clients/{clientId}/finding-dismissals/{id}": {
      "patch": {
        "tags": [
          "FindingDismissals"
        ],
        "summary": "Updates the label of an existing finding dismissal.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Dismissal identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Label update.",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFindingDismissalRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFindingDismissalRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFindingDismissalRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dismissal updated.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDismissalDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDismissalDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingDismissalDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid admin key.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Dismissal not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "FindingDismissals"
        ],
        "summary": "Deletes a finding dismissal.",
        "parameters": [
          {
            "name": "clientId",
            "in": "path",
            "description": "Client identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Dismissal identifier.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Dismissal deleted."
          },
          "401": {
            "description": "Missing or invalid admin key.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Dismissal not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "FindingDismissalDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "clientId": {
            "type": "string",
            "format": "uuid"
          },
          "patternText": {
            "type": "string"
          },
          "label": {
            "type": "string",
            "nullable": true
          },
          "originalMessage": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false,
        "description": "DTO for a finding dismissal."
      },
      "CreateFindingDismissalRequest": {
        "type": "object",
        "properties": {
          "originalMessage": {
            "type": "string"
          },
          "label": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request to create a finding dismissal."
      },
      "UpdateFindingDismissalRequest": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request to update a finding dismissal's label."
      },
...
    }
  },
...
}